Pull untwine 1.4.0 from upstream
