+++ external/untwine/epf/FileProcessor.cpp
@@ -134,11 +134,13 @@ void FileProcessor::run()
     opts.add("filename", m_fi.filename);
     opts.add("count", m_fi.numPoints);
     if (m_fi.driver == "readers.las")
+    {
         opts.add("nosrs", m_fi.no_srs);
+        opts.add("use_eb_vlr", "true");
 #ifdef PDAL_LAS_START
-    if (m_fi.driver == "readers.las")
         opts.add("start", m_fi.start);
 #endif
+    }
 
     pdal::StageFactory factory;
     pdal::Stage *s = factory.createStage(m_fi.driver);
