+++ external/untwine/epf/Reprocessor.cpp
@@ -19,6 +19,8 @@
 #include "Reprocessor.hpp"
 #include "../untwine/Common.hpp"
 
+#include <mapfile.hpp>  // untwine/os
+
 namespace untwine
 {
 namespace epf
@@ -51,7 +53,7 @@ Reprocessor::Reprocessor(const VoxelKey& k, int numPoints, int pointSize,
 
 void Reprocessor::run()
 {
-    auto ctx = mapFile(m_filename, true, 0, m_fileSize);
+    auto ctx = os::mapFile(m_filename, true, 0, m_fileSize);
     if (ctx.addr() == nullptr)
     {
         std::cerr << "FATAL: " + m_filename + ": " + ctx.what();
@@ -69,7 +71,7 @@ void Reprocessor::run()
         cell->advance();
         pos += m_pointSize;
     }
-    unmapFile(ctx);
+    os::unmapFile(ctx);
     pdal::FileUtils::deleteFile(m_filename);
 }
