+++ external/untwine/untwine/Untwine.cpp
@@ -23,6 +23,9 @@
 #include "../epf/Epf.hpp"
 #include "../bu/BuPyramid.hpp"
 
+#include <dirlist.hpp>    // untwine/os
+#include <stringconv.hpp> // untwine/os
+
 namespace untwine
 {
 
@@ -72,13 +75,21 @@ bool handleOptions(pdal::StringList& arglist, Options& options)
             std::cout << "untwine version (" << UNTWINE_VERSION << ")\n";
         if (help)
         {
-            std::cout << "Usage: untwine [output file/directory] <options>\n";
+            std::cout << "Usage: untwine output file <options>\n";
             programArgs.dump(std::cout, 2, 80);
         }
         if (help || version)
             return false;
 
         programArgs.parse(arglist);
+
+        // Make sure the output file can be opened so that we can provide an early error if
+        // there's a problem.
+        std::ofstream tmp(os::toNative(options.outputName), std::ios::out | std::ios::binary);
+        if (!tmp)
+            throw FatalError("Can't open file '" + options.outputName + "' for output");
+        tmp.close();
+        pdal::FileUtils::deleteFile(options.outputName);
 
         if (!tempArg->set())
         {
@@ -114,7 +125,7 @@ void cleanup(const std::string& dir, bool rmdir)
     std::regex re("[0-9]+-[0-9]+-[0-9]+-[0-9]+.bin");
     std::smatch sm;
 
-    const std::vector<std::string>& files = directoryList(dir);
+    const std::vector<std::string>& files = os::directoryList(dir);
     for (const std::string& f : files)
         if (std::regex_match(f, sm, re))
             pdal::FileUtils::deleteFile(dir + "/" + f);
@@ -136,7 +147,7 @@ int main(int argc, char *argv[])
     argv++;
     argc--;
     while (argc--)
-        arglist.push_back(untwine::fromNative(*argv++));
+        arglist.push_back(untwine::os::fromNative(*argv++));
 
     using namespace untwine;
