+++ external/untwine/untwine/generic/stringconv.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+namespace untwine
+{
+namespace os
+{
+
+inline std::string toNative(const std::string& in)
+{
+    return in;
+}
+
+inline std::string fromNative(const std::string& in)
+{
+    return in;
+}
+
+} // namespace os
+} // namespace untwine
